# public/openapi.yaml
openapi: 3.1.0
info:
  title: BasaltSurge API
  version: 1.0.0
  contact:
    name: BasaltSurge Support
    url: https://surge.basalthq.com
    email: support@basalthq.com
  description: >
    Public API surface for BasaltSurge under a hardened posture: Azure Front
    Door Premium (AFD) gateway → Azure API Management (APIM).

    All developer-facing endpoints require an APIM subscription key header
    `Ocp-Apim-Subscription-Key` and use the APIM custom domain. AFD is optional
    as a fallback.

    APIM policy permits requests with a valid subscription key and may also
    accept AFD-originated requests carrying `x-edge-secret` if AFD is
    configured. Direct backend origins are not supported.

    Wallet identity is resolved at the gateway based on subscription;
    client-supplied `x-wallet` header and `wallet` parameters are ignored.

    Health endpoint: `GET /healthz` returns 200 JSON for probes; does not
    require a subscription.

    Note: Some endpoints are read-only or environment-gated in production.
    Admin/sensitive routes (JWT-only) are not part of this spec.
  x-guidance: >
    Welcome to BasaltSurge's Agentic Commerce API.

    To use this API:

    1. Non-paid resources are fully accessible with standard API keys or without
    auth depending on the path.

    2. Payment-gated resources under /api/x402/ must be paid for. For example,
    POST /api/x402/orders to discover the pricing required. The server will
...

...

...
    the x402 challenge there.
...
servers:
...
security:
  - apimKey: []
...
paths:
...
  /api/x402/orders:
    post:
      summary: Create Order via x402 Payment
      operationId: x402CreateOrder
      tags:
        - Orders
      description: >
        x402-gated order endpoint for AI agents. Always returns 402 with payment
        challenge when no x-payment proof is attached. Agents discover shops via
        GET /api/directory/shops, resolve inventory, then POST here with
        shopSlug and items to receive a payment challenge. Once paid, a receipt
        is returned.
      security:
        - x402: []
      x-payment-info:
        price:
          mode: dynamic
          currency: USD
          min: '0.10'
          max: '100000.00'
        protocols:
          - x402: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - shopSlug
                - items
              properties:
                shopSlug:
                  type: string
                  description: The shop's unique slug (from /api/directory/shops)
                items:
                  type: array
                  description: Items to order with SKU and quantity
                  items:
                    type: object
                    required:
                      - sku
                      - qty
                    properties:
                      sku:
                        type: string
                      qty:
                        type: integer
                        minimum: 1
                jurisdictionCode:
                  type: string
                  description: Optional tax jurisdiction code (e.g. US-CA)
      responses:
        '200':
          description: Order receipt (payment settled)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  receipt:
                    type: object
                    properties:
                      receiptId:
                        type: string
                      totalUsd:
                        type: number
                      lineItems:
                        type: array
                        items:
                          type: object
                      status:
                        type: string
                        enum:
                          - paid
        '402':
          description: Payment Required – x402 challenge issued
          headers:
            Payment-Required:
              description: Base64 encoded payload
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  x402Version:
                    type: integer
                  error:
                    type: string
                  accepts:
                    type: array
                    items:
                      type: object
...
components:
  securitySchemes:
    apimKey:
      type: apiKey
      in: header
      name: Ocp-Apim-Subscription-Key
      description: >
        Azure API Management subscription key. Required for developer-facing
        endpoints.

        Obtain your key via BasaltSurge Admin → API Subscriptions (or directly
        in APIM).
    x402:
      type: apiKey
      in: header
      name: x-payment
      description: x402 Agentic Payment Standard Proof (Base64 payload)
...